--- packages/ocom/service-blob-storage/src/feature-flags.local.ts
@@ -1,1 +1,0 @@
-export const FeatureFlagsLocal = JSON.stringify({ FeatureFlags: [] });
