+++ packages/ocom/service-blob-storage/src/feature-flags.schema.json
@@ -1,0 +1,40 @@
+{
+	"$schema": "https://json-schema.org/draft/2020-12/schema",
+	"type": "object",
+	"title": "Feature Flags",
+	"properties": {
+		"FeatureFlags": {
+			"type": "array",
+			"items": {
+				"type": "object",
+				"properties": {
+					"Name": {
+						"type": "string",
+						"examples": ["MAINTENANCE_UPCOMING_UI_COMMUNITY_PORTAL"]
+					},
+					"Description": {
+						"type": "string",
+						"examples": ["Indicates whether scheduled maintenance is upcoming for the Community Portal"]
+					},
+					"Value": {
+						"type": "string",
+						"examples": ["false"]
+					},
+					"AllowedValues": {
+						"type": "array",
+						"items": {
+							"type": "string",
+							"examples": ["true"]
+						}
+					},
+					"RetirementDate": {
+						"type": "string",
+						"examples": ["2026-01-09"]
+					}
+				},
+				"additionalProperties": false
+			}
+		}
+	},
+	"additionalProperties": false
+}
