+++ packages/ocom/service-blob-storage/src/index.ts
@@ -9,6 +9,6 @@ export type {
 	UploadTextBlobRequest,
 } from '@cellix/service-blob-storage';
 export type { BlobStorageOperations, ClientUploadOperations, CreateBlobAccessUrlRequest } from './blob-storage.contract.ts';
-export type { FeatureFlag, FeatureFlagsPayloadType } from './feature-flags.ts';
+export type { FeatureFlag, FeatureFlagsPayloadType } from './feature-flags.payload-type.ts';
 export { ServiceBlobStorage } from './service-blob-storage.ts';
 export { ServiceClientBlobStorage } from './service-client-blob-storage.ts';
